记录添加 记录删除 记录修改分块处理 读取数据精度设定（未测试版）

+++ src/balance_calculate.h
@@ -6,9 +6,9 @@
 #define SALARYMANAGESYSTEM_BALANCE_CALCULATE_H
 
 
-float disburseCalculate(int row, int count);
+double disburseCalculate(int row, int count);
 
-float incomeCalculate(int row, int count);
+double incomeCalculate(int row, int count);
 
 void balance_change(int row);
